+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Description: Kinova Gen3 Lite (URDF)
 - Description: ARX L5 (MJCF) (thanks to @jonzamora)
+- Description: Kinova Gen3 Lite (URDF)
+- Description: Low-cost robot arm (MJCF)
 
 ### Changed